[add] ✨スコアとコンボ作った
一旦1秒間AddScoreが起きなかったらリセットにしてます
使うかどうかわからないけどUIに起こせるようにもしました

+++ Assets/Scripts/Main/Enemy/EnemyData.cs
@@ -9,5 +9,5 @@ public class EnemyData : ScriptableObject
     public float enemyHP = 20f; //敵のHP
     public float enemyDamage = 20f; //敵の攻撃力、エネミーショットのダメージ
     public float enemyHPRate = 10.0f; // エネミーショットで回復する倍率
-    public float enemyScore = 100.0f; // 敵のスコア
+    public int enemyScore = 100; // 敵のスコア
 }
